+++ Modules/Identity/Infrastructure/Persistence/Repositories/AddressRepositoryInterface.php
@@ -27,5 +27,7 @@ interface AddressRepositoryInterface
 
     public function findById(int $id): ?Address;
 
+    public function findByPublicCode(string $publicCode): ?Address;
+
     public function refreshWithRelations(Address $address): Address;
 }